requirements.txt
numpy
numpy-stl
plotly
pypotrace